// client/src/actions/types.ts
enum Action {
	AUTH_LOGIN = "AUTH_LOGIN",
	FETCH_ENUMS = "FETCH_ENUMS",
	FETCH_USERS = "FETCH_USERS",
	FETCH_VEHICLES = "FETCH_VEHICLES",
	FETCH_BOOKINGS = "FETCH_BOOKINGS",
	FETCH_CURRENT_USER_DETAILS = "FETCH_CURRENT_USER_DETAILS",
	AUTH_LOGOUT = "AUTH_LOGOUT",
	FETCH_LOCATIONS = "FETCH_LOCATIONS",
	FETCH_ACCIDENTS = "FETCH_ACCIDENTS",
	FETCH_CATEGORIES = "FETCH_CATEGORIES"
}

export default Action;